Add `tmuxify.sh` script
Opens a project in a tmux session.

Pipes results of `fd` for project directories to `FZF`. Takes the result and
either creates a new tmux session (with the name of the project) or attaches
to a session if one already exists for that project.

The inspiration for this came from @ThePrimeagen's [Developer Productivity][1]
course.

[1]: https://frontendmasters.com/courses/developer-productivity/